pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "scapr"
dynamic = ["version", "description"]
authors = [{name = "Bill Melvin", email = "bill@billmelvin.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
keywords = ["image", "screen capture", "screen shot", "screenshot"]
requires-python = ">=3.7"
dependencies = ["pillow"]

[project.urls]
Home = "https://github.com/wmelvin/scapr"

[project.optional-dependencies]
test = ["pytest"]
